Bump version to 1.7.2 and improve build script for versioning

+++ release/pyi_win.ps1
@@ -9,16 +9,21 @@ if ($git_commit -eq $null) {
 }
 
 # delete the previous build
-Remove-Item -Recurse -Force .\release\EasyFlowQ_release_win\
+if (Test-Path .\release\EasyFlowQ_release_win\) {
+    Remove-Item -Recurse -Force .\release\EasyFlowQ_release_win\
+}
 
 # build the app
-$versionString = "$_ver" + "_bld_" + $git_commit
+$versionString = $_ver + "_bld_" + $git_commit
+echo "Building EasyFlowQ with pyinstaller, version: $versionString"
+
 $pyinstallerArgs = @(
     "--noconfirm"
     "--distpath", ".\release\EasyFlowQ_release_win\"
     "--workpath", ".\_temp\"
     ".\release\pyi_universal.spec"
-    "--version", $versionString
+    "--"
+    "--version", $_ver
 )
 pyinstaller @pyinstallerArgs
 
@@ -26,5 +31,8 @@ pyinstaller @pyinstallerArgs
 Start-Sleep -Seconds 5
 
 # compress the build
-Compress-Archive -Force -Path .\release\EasyFlowQ_release_win\EasyFlowQ_Bundle_v$versionString `
+Compress-Archive -Force -Path .\release\EasyFlowQ_release_win\EasyFlowQ_Bundle_v$_ver `
     -DestinationPath .\release\EasyFlowQ_release_win\EasyFlowQ_Win_v$versionString.zip
+
+# start the easfylowq app
+Start-Process -FilePath .\release\EasyFlowQ_release_win\EasyFlowQ_Bundle_v$_ver\EasyFlowQ_v$_ver.exe
